api/requirements.txt
asgiref==3.2.7
astroid==2.3.2
Django==3.0.5
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
wrapt==1.11.2
